fix: #87 as suggested from @mitra42 this now should be fixed and code formatted

+++ src/protocols/lora_mesh/services/superframe_service.cpp
@@ -543,8 +543,8 @@ SuperframeService::SuperframeStats SuperframeService::GetSuperframeStats()
     if (is_running_) {
         uint32_t current_time = GetRTOS().getTickCount();
         stats.total_runtime_ms = current_time - service_start_time_;
-        stats.sync_drift_ms =
-            sync_drift_accumulator_ / std::max(1u, superframes_completed_);
+        stats.sync_drift_ms = sync_drift_accumulator_ /
+                              std::max((uint32_t)1u, superframes_completed_);
     }
 
     if (timing_samples_ > 0) {
